requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
django-environ==0.13.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg==3.3.3
psycopg2-binary==2.9.11
PyJWT==2.11.0
sqlparse==0.5.5
typing_extensions==4.15.0
python-dotenv==1.0.1
gunicorn==21.2.0
